Refactor public tabular ingest hooks
